fluent4rs
Module walker

The Walker enables the Resource returned by the Parser to be traversed. This functionality requires the walker feature.

The default Visitor implementation is a no-op, but selecting the trace feature prints the AST to stderr as the tree is traverse. Selecting the trace feature will also include the walker feature.

Structs

Walker

A Walker enables an AST node to be walked with the Walk::walk(node, visitor) function.

Traits

Visitor

Visitor enables key AST nodes to be visited. Implement Visitor and override required methods to allow them to be visited during a walk of the AST. If there’s a build with the trace feature, the default Visitor will output each node to stderr.

Walkable

AST nodes implement walkable, so any can be re-walked and / ot re-visited. It is not expected that this trait will be implemented by an end user.