fluent4rs::walker
Trait Walkable

pub trait Walkable {
    // Required method
    fn walk(&self, visitor: &mut dyn Visitor);
}

AST nodes implement walkable, so any can be re-walked and / ot re-visited. It is not expected that this trait will be implemented by an end user.

Required Methods

fn walk(&self, visitor: &mut dyn Visitor)